// S13_2/product-composite-app/src/app/api/model/productAggregateDto.ts
/**
 * API Product-Composite
 *
 * Contact: soporte@miempresa.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { RecommendationSummaryDto } from './recommendationSummaryDto';
import { ServiceAddressesDto } from './serviceAddressesDto';
import { ReviewSummaryDto } from './reviewSummaryDto';


export interface ProductAggregateDto { 
    /**
     * Identificador del producto
     */
    productId?: number;
    /**
     * Nombre del producto
     */
    name: string;
    /**
     * Peso del producto
     */
    weight: number;
    /**
     * Lista de recomendaciones del producto
     */
    recommendations?: Array<RecommendationSummaryDto>;
    /**
     * Lista de reseñas del producto
     */
    reviews?: Array<ReviewSummaryDto>;
    serviceAddresses?: ServiceAddressesDto;
}